backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
django-redis==5.4.0
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-storages==1.14.2
django-cors-headers==4.3.1
python-decouple==3.8
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0